docs :  Add new folder to separate images
