+++ src/assets/fonts/font.css
@@ -10,5 +10,5 @@
 
 @font-face{
     font-family : "NanumSquareRound";
-    src : url('./NanumSquareRoundOTFEB.otf')
+    src : url('./NanumSquareRoundOTFB.otf')
 }
