+++ src/component/Banner/Banner.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react'
 import styled from 'styled-components'
 import {PageHeader,MenuWrapper,MainMenu,MainMenuWrapper,LoginMenuWrapper,LoginMenu} from'./BannerStyle'
-import {ReactComponent as HangangLogo} from '../../assets/images/HangangLogo.svg'
+import {ReactComponent as HangangLogo} from '../../assets/images/banner/HangangLogo.svg'
 const menuList = ['홈','강의평','강의자료','시간표']
 const DevideLine = styled.div`
     width : 1px;
